+++ RiotSwiftUI/Modules/Authentication/Common/Service/MatrixSDK/AuthenticationService.swift
@@ -174,13 +174,14 @@ class AuthenticationService: NSObject {
     }
     
     /// Reset the service to a fresh state.
-    func reset() {
+    /// - Parameter useDefaultServer: Pass `true` to revert back to the one in `BuildSettings`, otherwise the current homeserver will be kept.
+    func reset(useDefaultServer: Bool = false) {
         loginWizard = nil
         registrationWizard = nil
         softLogoutCredentials = nil
 
         // The previously used homeserver is re-used as `startFlow` will be called again a replace it anyway.
-        let address = state.homeserver.addressFromUser ?? state.homeserver.address
+        let address = useDefaultServer ? BuildSettings.serverConfigDefaultHomeserverUrlString : state.homeserver.addressFromUser ?? state.homeserver.address
         let identityServer = state.identityServer
         self.state = AuthenticationState(flow: .login,
                                          homeserverAddress: address,
@@ -195,27 +196,6 @@ class AuthenticationService: NSObject {
     func continueSSOLogin(with token: String, and transactionID: String) -> Bool {
         delegate?.authenticationService(self, didReceive: token, with: transactionID) ?? false
     }
-    
-//    /// Perform a well-known request, using the domain from the matrixId
-//    func getWellKnownData(matrixId: String,
-//                          homeServerConnectionConfig: HomeServerConnectionConfig?) async -> WellknownResult {
-//
-//    }
-//
-//    /// Authenticate with a matrixId and a password
-//    /// Usually call this after a successful call to getWellKnownData()
-//    /// - Parameter homeServerConnectionConfig the information about the homeserver and other configuration
-//    /// - Parameter matrixId the matrixId of the user
-//    /// - Parameter password the password of the account
-//    /// - Parameter initialDeviceName the initial device name
-//    /// - Parameter deviceId the device id, optional. If not provided or null, the server will generate one.
-//    func directAuthentication(homeServerConnectionConfig: HomeServerConnectionConfig,
-//                              matrixId: String,
-//                              password: String,
-//                              initialDeviceName: String,
-//                              deviceId: String? = nil) async -> MXSession {
-//        
-//    }
     
     // MARK: - Private