+++ RiotSwiftUI/Modules/Authentication/Registration/AuthenticationRegistrationViewModel.swift
@@ -48,8 +48,8 @@ class AuthenticationRegistrationViewModel: AuthenticationRegistrationViewModelTy
             Task { await validateUsername() }
         case .enablePasswordValidation:
             Task { await enablePasswordValidation() }
-        case .clearUsernameError:
-            Task { await clearUsernameError() }
+        case .resetUsernameAvailability:
+            Task { await resetUsernameAvailability() }
         case .next:
             Task { await callback?(.createAccount(username: state.bindings.username, password: state.bindings.password)) }
         case .continueWithSSO(let provider):
@@ -59,14 +59,29 @@ class AuthenticationRegistrationViewModel: AuthenticationRegistrationViewModelTy
         }
     }
     
+    @MainActor func update(isLoading: Bool) {
+        guard state.isLoading != isLoading else { return }
+        state.isLoading = isLoading
+    }
+    
     @MainActor func update(homeserver: AuthenticationHomeserverViewData) {
         state.homeserver = homeserver
+    }
+    
+    @MainActor func update(username: String) {
+        guard username != state.bindings.username else { return }
+        state.bindings.username = username
+    }
+    
+    @MainActor func confirmUsernameAvailability(_ username: String) {
+        guard username == state.bindings.username else { return }
+        state.usernameAvailability = .available
     }
     
     @MainActor func displayError(_ type: AuthenticationRegistrationErrorType) {
         switch type {
         case .usernameUnavailable(let message):
-            state.usernameErrorMessage = message
+            state.usernameAvailability = .invalid(message)
         case .mxError(let message):
             state.bindings.alertInfo = AlertInfo(id: type,
                                                  title: VectorL10n.error,
@@ -101,9 +116,9 @@ class AuthenticationRegistrationViewModel: AuthenticationRegistrationViewModelTy
         state.hasEditedPassword = true
     }
     
-    /// Clear any errors being shown in the username text field footer.
-    @MainActor private func clearUsernameError() {
-        guard state.usernameErrorMessage != nil else { return }
-        state.usernameErrorMessage = nil
+    /// Reset the username's availability, clearing any messages being shown in the username text field footer.
+    @MainActor private func resetUsernameAvailability() {
+        if case .unknown = state.usernameAvailability { return }
+        state.usernameAvailability = .unknown
     }
 }