+++ RiotSwiftUI/Modules/Authentication/SoftLogout/Coordinator/AuthenticationSoftLogoutCoordinator.swift
@@ -155,7 +155,8 @@ final class AuthenticationSoftLogoutCoordinator: Coordinator, Presentable {
         let modalRouter = NavigationRouter()
 
         let parameters = AuthenticationForgotPasswordCoordinatorParameters(navigationRouter: modalRouter,
-                                                                           loginWizard: loginWizard)
+                                                                           loginWizard: loginWizard,
+                                                                           homeserver: parameters.authenticationService.state.homeserver)
         let coordinator = AuthenticationForgotPasswordCoordinator(parameters: parameters)
         coordinator.callback = { [weak self, weak coordinator] result in
             guard let self = self, let coordinator = coordinator else { return }