+++ RiotSwiftUI/Modules/Common/Mock/ScreenList.swift
@@ -18,33 +18,58 @@ import SwiftUI
 
 struct ScreenList: View {
     
-    private var allStates: [ScreenStateInfo]
+    private let allStates: [ScreenStateInfo]
+    
+    @State private var searchQuery = ""
+    @State private var filteredStates: [ScreenStateInfo]
     
     init(screens: [MockScreenState.Type]) {
-        allStates = screens
-            .map({ $0.stateRenderer })
-            .flatMap{( $0.states )}
+        let states = screens
+            .map { $0.stateRenderer }
+            .flatMap { $0.states }
+        
+        allStates = states
+        filteredStates = states
     }
     
     var body: some View {
         NavigationView {
-            List {
-                SwiftUI.Section {
-                    ForEach(0..<allStates.count, id: \.self) { i in
-                        let state = allStates[i]
-                        NavigationLink(destination: state.view) {
-                            Text(state.screenTitle)
+            VStack {
+                TextField("Search", text: $searchQuery)
+                    .textFieldStyle(.roundedBorder)
+                    .padding(.horizontal)
+                    .accessibilityIdentifier("searchQueryTextField")
+                    .onChange(of: searchQuery, perform: search)
+                
+                Form {
+                    SwiftUI.Section {
+                        ForEach(0..<filteredStates.count, id: \.self) { i in
+                            let state = filteredStates[i]
+                            NavigationLink(destination: state.view) {
+                                Text(state.screenTitle)
+                            }
                         }
+                    } footer: {
+                        Text("End of list")
+                            .accessibilityIdentifier("footerText")
                     }
                 }
-                
-                SwiftUI.Section {
-                    Text("Last Item")
-                        .accessibilityIdentifier("lastItem")
-                }
             }
+            .background(Color(.systemGroupedBackground).ignoresSafeArea())
+            .navigationTitle("SwiftUI Screens")
+            .navigationBarTitleDisplayMode(.inline)
         }
         .navigationTitle("Screen States")
+    }
+    
+    func search(query: String) {
+        if query.isEmpty {
+            filteredStates = allStates
+        } else {
+            filteredStates = allStates.filter {
+                $0.screenTitle.localizedStandardContains(query)
+            }
+        }
     }
 }