+++ RiotSwiftUI/Modules/Room/TimelinePoll/Test/UI/TimelinePollUITests.swift
@@ -17,17 +17,7 @@
 import XCTest
 import RiotSwiftUI
 
-class TimelinePollUITests: XCTestCase {
-    
-    private var app: XCUIApplication!
-    
-    override func setUp() {
-        continueAfterFailure = false
-                
-        app = XCUIApplication()
-        app.launch()
-    }
-    
+class TimelinePollUITests: MockScreenTestCase {
     func testOpenDisclosedPoll() {
         app.goToScreenWithIdentifier(MockTimelinePollScreenState.openDisclosed.title)